% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{as.data.frame.dynarules}
\alias{as.data.frame.dynarules}
\title{Coerce Mined Rules to a Data Frame}
\usage{
\method{as.data.frame}{dynarules}(x, ...)
}
\arguments{
\item{x}{A \code{dynarules} object.}

\item{...}{Ignored.}
}
\value{
The tidy rules data.frame.
}
\description{
Coerce Mined Rules to a Data Frame
}
