% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{print.dyna_transactions}
\alias{print.dyna_transactions}
\title{Print Method for dyna_transactions}
\usage{
\method{print}{dyna_transactions}(x, ...)
}
\arguments{
\item{x}{A \code{dyna_transactions} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for dyna_transactions
}
