% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dynarules}
\alias{print.dynarules}
\title{Print Method for dynarules}
\usage{
\method{print}{dynarules}(x, ...)
}
\arguments{
\item{x}{A \code{dynarules} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for dynarules
}
