% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dynarules_boot}
\alias{print.dynarules_boot}
\title{Print Method for Bootstrapped Rules}
\usage{
\method{print}{dynarules_boot}(x, ...)
}
\arguments{
\item{x}{A \code{dynarules_boot} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for Bootstrapped Rules
}
