% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{print.dynarules_cba}
\alias{print.dynarules_cba}
\title{Print Method for a CBA Classifier}
\usage{
\method{print}{dynarules_cba}(x, ...)
}
\arguments{
\item{x}{A \code{dynarules_cba} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for a CBA Classifier
}
