% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{print.dynarules_net}
\alias{print.dynarules_net}
\title{Print Method for dynarules_net}
\usage{
\method{print}{dynarules_net}(x, ...)
}
\arguments{
\item{x}{A \code{dynarules_net} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for dynarules_net
}
