% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dynarules_perm}
\alias{print.dynarules_perm}
\title{Print Method for dynarules_perm}
\usage{
\method{print}{dynarules_perm}(x, ...)
}
\arguments{
\item{x}{A \code{dynarules_perm} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for dynarules_perm
}
