% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dynarules}
\alias{summary.dynarules}
\title{Summary Method for dynarules}
\usage{
\method{summary}{dynarules}(object, ...)
}
\arguments{
\item{object}{A \code{dynarules} object.}

\item{...}{Ignored.}
}
\value{
The tidy rules data.frame.
}
\description{
Summary Method for dynarules
}
