% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dynarules_boot}
\alias{summary.dynarules_boot}
\title{Summary Method for dynarules_boot}
\usage{
\method{summary}{dynarules_boot}(object, ...)
}
\arguments{
\item{object}{A \code{dynarules_boot} object.}

\item{...}{Ignored.}
}
\value{
The tidy stability data.frame.
}
\description{
Summary Method for dynarules_boot
}
