% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dynarules_perm}
\alias{summary.dynarules_perm}
\title{Summary Method for dynarules_perm}
\usage{
\method{summary}{dynarules_perm}(object, ...)
}
\arguments{
\item{object}{A \code{dynarules_perm} object.}

\item{...}{Ignored.}
}
\value{
The tidy tests data.frame.
}
\description{
Summary Method for dynarules_perm
}
